split install up into smaller functions
