+++ pkg/install/graph.go
@@ -4,10 +4,13 @@ package install
 // Licensed under the Apache License 2.0.
 
 import (
+	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
 	"reflect"
 
+	azstorage "github.com/Azure/azure-sdk-for-go/storage"
 	"github.com/openshift/installer/pkg/asset"
 	"github.com/openshift/installer/pkg/asset/cluster"
 	"github.com/openshift/installer/pkg/asset/ignition/bootstrap"
@@ -19,9 +22,11 @@ import (
 	"github.com/openshift/installer/pkg/asset/password"
 	"github.com/openshift/installer/pkg/asset/releaseimage"
 	"github.com/openshift/installer/pkg/asset/rhcos"
+	"github.com/openshift/installer/pkg/asset/targets"
 	"github.com/openshift/installer/pkg/asset/templates/content/bootkube"
 	"github.com/openshift/installer/pkg/asset/templates/content/openshift"
 	"github.com/openshift/installer/pkg/asset/tls"
+	uuid "github.com/satori/go.uuid"
 )
 
 var registeredTypes = map[string]asset.Asset{
@@ -124,10 +129,82 @@ var registeredTypes = map[string]asset.Asset{
 
 type graph map[reflect.Type]asset.Asset
 
-func (g graph) resolve(a asset.Asset) (asset.Asset, error) {
+type Graph interface {
+	Store(blobService *azstorage.BlobStorageClient) error
+	Resolve(a asset.Asset) (asset.Asset, error)
+	UnmarshalJSON(b []byte) error
+	MarshalJSON() ([]byte, error)
+	GetMap() map[reflect.Type]asset.Asset
+}
+
+func NewGraph(ctx context.Context, installConfig *installconfig.InstallConfig, platformCreds *installconfig.PlatformCreds, image *releaseimage.Image) (Graph, error) {
+	clusterID := &installconfig.ClusterID{
+		UUID:    uuid.NewV4().String(),
+		InfraID: "aro",
+	}
+
+	g := graph{
+		reflect.TypeOf(installConfig): installConfig,
+		reflect.TypeOf(platformCreds): platformCreds,
+		reflect.TypeOf(image):         image,
+		reflect.TypeOf(clusterID):     clusterID,
+	}
+
+	for _, a := range targets.Cluster {
+		_, err := g.Resolve(a)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return &g, nil
+}
+
+func LoadGraph(blobService *azstorage.BlobStorageClient) (Graph, error) {
+	aro := blobService.GetContainerReference("aro")
+	cluster := aro.GetBlobReference("graph")
+	rc, err := cluster.Get(nil)
+	if err != nil {
+		return nil, err
+	}
+	defer rc.Close()
+
+	var g graph
+	err = json.NewDecoder(rc).Decode(&g)
+	if err != nil {
+		return nil, err
+	}
+
+	return &g, nil
+}
+
+func (g graph) GetMap() map[reflect.Type]asset.Asset {
+	return g
+}
+
+func (g graph) Store(blobService *azstorage.BlobStorageClient) error {
+	bootstrap := g[reflect.TypeOf(&bootstrap.Bootstrap{})].(*bootstrap.Bootstrap)
+
+	bootstrapIgn := blobService.GetContainerReference("ignition").GetBlobReference("bootstrap.ign")
+	err := bootstrapIgn.CreateBlockBlobFromReader(bytes.NewReader(bootstrap.File.Data), nil)
+	if err != nil {
+		return err
+	}
+
+	// the graph is quite big so we store it in a storage account instead of
+	// in cosmosdb
+	graph := blobService.GetContainerReference("aro").GetBlobReference("graph")
+	b, err := json.MarshalIndent(g, "", "    ")
+	if err != nil {
+		return err
+	}
+
+	return graph.CreateBlockBlobFromReader(bytes.NewReader(b), nil)
+}
+
+func (g graph) Resolve(a asset.Asset) (asset.Asset, error) {
 	if _, found := g[reflect.TypeOf(a)]; !found {
 		for _, dep := range a.Dependencies() {
-			_, err := g.resolve(dep)
+			_, err := g.Resolve(dep)
 			if err != nil {
 				return nil, err
 			}
